// src/app/api/model/groupDtoIEnumerableApiResponse.ts
/**
 * TaskFlow Chat API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupDto } from './groupDto';


export interface GroupDtoIEnumerableApiResponse { 
    success?: boolean;
    data?: Array<GroupDto> | null;
    message?: string | null;
    errorCode?: string | null;
    details?: any | null;
    traceId?: string | null;
}