// src/app/api/model/messageDtoIEnumerableApiResponse.ts
/**
 * TaskFlow Chat API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessageDto } from './messageDto';


export interface MessageDtoIEnumerableApiResponse { 
    success?: boolean;
    data?: Array<MessageDto> | null;
    message?: string | null;
    errorCode?: string | null;
    details?: any | null;
    traceId?: string | null;
}